+++ pro/src/hooks/__specs__/useNotification.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import useNotification from 'hooks/useNotification'
-import * as notificationReducer from 'store/reducers/notificationReducer'
+import * as notificationReducer from 'store/notifications/reducer'
 import { renderWithProviders } from 'utils/renderWithProviders'
 
 const TestComponent = (): JSX.Element | null => {
