+++ pro/src/store/reducers/index.ts
@@ -1,16 +1,16 @@
 import { combineReducers } from 'redux'
 
 import { featuresReducer } from 'store/features/reducer'
+import { notificationsReducer } from 'store/notifications/reducer'
 import { offersReducer } from 'store/offers/reducer'
 import { userReducer } from 'store/user/reducer'
 
 import maintenanceReducer from './maintenanceReducer'
-import { notificationReducer } from './notificationReducer'
 
 const rootReducer = combineReducers({
   features: featuresReducer,
   offers: offersReducer,
-  notification: notificationReducer,
+  notification: notificationsReducer,
   user: userReducer,
   maintenance: maintenanceReducer,
 })
