+++ pro/src/store/testUtils.ts
@@ -1,9 +1,9 @@
 import { FeatureResponseModel, UserRole } from 'apiClient/v1'
 import createStore from 'store'
 import { initialState as featuresInitialState } from 'store/features/reducer'
+import { initialState as notificationInitialState } from 'store/notifications/reducer'
 import { initialState as offersInitialState } from 'store/offers/reducer'
 import { initialState as maintenanceInitialState } from 'store/reducers/maintenanceReducer'
-import { initialState as notificationInitialState } from 'store/reducers/notificationReducer'
 import { initialState as userInitialState } from 'store/user/reducer'
 
 import { RootState } from './reducers'
